+++ frontend/v1/domain/services/IDatasetStorage.ts
@@ -1,4 +1,4 @@
-import { Dataset } from "../entities/Dataset";
+import { Dataset } from "../entities/dataset/Dataset";
 
 export interface IDatasetStorage {
   save(dataset: Dataset): void;