+++ frontend/v1/infrastructure/storage/DatasetStorage.ts
@@ -1,4 +1,4 @@
-import { Dataset } from "@/v1/domain/entities/Dataset";
+import { Dataset } from "~/v1/domain/entities/dataset/Dataset";
 import { IDatasetStorage } from "@/v1/domain/services/IDatasetStorage";
 import { useStoreFor } from "@/v1/store/create";